notificaciones/requirements.txt
redis>=5.0.0
resend>=2.0.0
python-dotenv>=1.0.0
